# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THÚ GIỚI AI - World Creature Simulation API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Creature {
  id: number;
  name: string;
  rank: string;
  rankLevel: number;
  element: string;
  habitat: string;
  population: number;
  lifespan: number;
  reproductionRate: number;
  description: string;
  status: string;
  isHybrid: boolean;
  /** @nullable */
  parentA?: string | null;
  /** @nullable */
  parentB?: string | null;
  createdAt: string;
}

export interface Zone {
  id: number;
  name: string;
  temperature: number;
  resources: string;
  capacity: number;
  currentPopulation: number;
  description?: string;
}

export interface ZoneDetail {
  id: number;
  name: string;
  temperature: number;
  resources: string;
  capacity: number;
  currentPopulation: number;
  description?: string;
  creatures: Creature[];
}

export interface BreedingInput {
  creatureAId: number;
  creatureBId: number;
}

export interface BreedingResult {
  newCreature: Creature;
  message: string;
}

export interface HistoryEntry {
  id: number;
  worldDay: number;
  eventType: string;
  description: string;
  createdAt: string;
}

export interface ZoneStat {
  zoneId: number;
  zoneName: string;
  population: number;
  speciesCount: number;
  capacity: number;
}

export interface DashboardSummary {
  totalSpecies: number;
  totalPopulation: number;
  extinctSpecies: number;
  livingSpecies: number;
  worldDay: number;
  zoneStats: ZoneStat[];
  recentEvents: HistoryEntry[];
}

export interface CreaturePopulation {
  id: number;
  name: string;
  population: number;
  status: string;
}

export interface TickResult {
  worldDay: number;
  events: string[];
  populations: CreaturePopulation[];
}

export type ListCreaturesParams = {
zone?: string;
element?: string;
status?: string;
};

export type ListHistoryParams = {
limit?: number;
offset?: number;
};