# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: THÚ GIỚI AI - World Creature Simulation API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: creatures
    description: Creature management
  - name: zones
    description: Zone management
  - name: breeding
    description: Breeding operations
  - name: history
    description: World history log
  - name: dashboard
    description: Dashboard summary
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get world dashboard summary
      description: Returns total population, species count, extinction count, and per-zone stats
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /creatures:
    get:
      operationId: listCreatures
      tags: [creatures]
      summary: List all creature species
      parameters:
        - name: zone
          in: query
          schema:
            type: string
        - name: element
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of creatures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Creature"

  /creatures/{id}:
    get:
      operationId: getCreature
      tags: [creatures]
      summary: Get creature detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Creature detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creature"
        "404":
          description: Not found

  /zones:
    get:
      operationId: listZones
      tags: [zones]
      summary: List all zones with current populations
      responses:
        "200":
          description: List of zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Zone"

  /zones/{id}:
    get:
      operationId: getZone
      tags: [zones]
      summary: Get zone detail with creatures
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Zone detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneDetail"
        "404":
          description: Not found

  /breeding:
    post:
      operationId: breedCreatures
      tags: [breeding]
      summary: Breed two creatures to create a new species
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BreedingInput"
      responses:
        "201":
          description: New creature born
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BreedingResult"
        "400":
          description: Invalid breeding pair

  /history:
    get:
      operationId: listHistory
      tags: [history]
      summary: Get world history logs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: History log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoryEntry"

  /simulation/tick:
    post:
      operationId: simulationTick
      tags: [dashboard]
      summary: Manually trigger a simulation tick (world engine step)
      responses:
        "200":
          description: Tick result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TickResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Creature:
      type: object
      required: [id, name, rank, rankLevel, element, habitat, population, lifespan, reproductionRate, description, status, isHybrid, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        rank:
          type: string
        rankLevel:
          type: integer
        element:
          type: string
        habitat:
          type: string
        population:
          type: integer
        lifespan:
          type: integer
        reproductionRate:
          type: number
        description:
          type: string
        status:
          type: string
        isHybrid:
          type: boolean
        parentA:
          type: ["string", "null"]
        parentB:
          type: ["string", "null"]
        createdAt:
          type: string

    Zone:
      type: object
      required: [id, name, temperature, resources, capacity, currentPopulation]
      properties:
        id:
          type: integer
        name:
          type: string
        temperature:
          type: number
        resources:
          type: string
        capacity:
          type: integer
        currentPopulation:
          type: integer
        description:
          type: string

    ZoneDetail:
      type: object
      required: [id, name, temperature, resources, capacity, currentPopulation, creatures]
      properties:
        id:
          type: integer
        name:
          type: string
        temperature:
          type: number
        resources:
          type: string
        capacity:
          type: integer
        currentPopulation:
          type: integer
        description:
          type: string
        creatures:
          type: array
          items:
            $ref: "#/components/schemas/Creature"

    BreedingInput:
      type: object
      required: [creatureAId, creatureBId]
      properties:
        creatureAId:
          type: integer
        creatureBId:
          type: integer

    BreedingResult:
      type: object
      required: [newCreature, message]
      properties:
        newCreature:
          $ref: "#/components/schemas/Creature"
        message:
          type: string

    HistoryEntry:
      type: object
      required: [id, worldDay, eventType, description, createdAt]
      properties:
        id:
          type: integer
        worldDay:
          type: integer
        eventType:
          type: string
        description:
          type: string
        createdAt:
          type: string

    DashboardSummary:
      type: object
      required: [totalSpecies, totalPopulation, extinctSpecies, livingSpecies, worldDay, zoneStats, recentEvents]
      properties:
        totalSpecies:
          type: integer
        totalPopulation:
          type: integer
        extinctSpecies:
          type: integer
        livingSpecies:
          type: integer
        worldDay:
          type: integer
        zoneStats:
          type: array
          items:
            $ref: "#/components/schemas/ZoneStat"
        recentEvents:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"

    ZoneStat:
      type: object
      required: [zoneId, zoneName, population, speciesCount, capacity]
      properties:
        zoneId:
          type: integer
        zoneName:
          type: string
        population:
          type: integer
        speciesCount:
          type: integer
        capacity:
          type: integer

    TickResult:
      type: object
      required: [worldDay, events, populations]
      properties:
        worldDay:
          type: integer
        events:
          type: array
          items:
            type: string
        populations:
          type: array
          items:
            $ref: "#/components/schemas/CreaturePopulation"

    CreaturePopulation:
      type: object
      required: [id, name, population, status]
      properties:
        id:
          type: integer
        name:
          type: string
        population:
          type: integer
        status:
          type: string