# lib/api-zod/src/generated/types/breedingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THÚ GIỚI AI - World Creature Simulation API
 * OpenAPI spec version: 0.1.0
 */

export interface BreedingInput {
  creatureAId: number;
  creatureBId: number;
}