# lib/api-zod/src/generated/types/breedingResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THÚ GIỚI AI - World Creature Simulation API
 * OpenAPI spec version: 0.1.0
 */
import type { Creature } from './creature';

export interface BreedingResult {
  newCreature: Creature;
  message: string;
}