# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THÚ GIỚI AI - World Creature Simulation API
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryEntry } from './historyEntry';
import type { ZoneStat } from './zoneStat';

export interface DashboardSummary {
  totalSpecies: number;
  totalPopulation: number;
  extinctSpecies: number;
  livingSpecies: number;
  worldDay: number;
  zoneStats: ZoneStat[];
  recentEvents: HistoryEntry[];
}