# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THÚ GIỚI AI - World Creature Simulation API
 * OpenAPI spec version: 0.1.0
 */

export interface HistoryEntry {
  id: number;
  worldDay: number;
  eventType: string;
  description: string;
  createdAt: string;
}