# lib/api-zod/src/generated/types/tickResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THÚ GIỚI AI - World Creature Simulation API
 * OpenAPI spec version: 0.1.0
 */
import type { CreaturePopulation } from './creaturePopulation';

export interface TickResult {
  worldDay: number;
  events: string[];
  populations: CreaturePopulation[];
}