# lib/api-zod/src/generated/types/zoneStat.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * THÚ GIỚI AI - World Creature Simulation API
 * OpenAPI spec version: 0.1.0
 */

export interface ZoneStat {
  zoneId: number;
  zoneName: string;
  population: number;
  speciesCount: number;
  capacity: number;
}